2.1.4
Handle not having a briefcase in the relay script, if for some reason
you visit that URL. Don’t display the enchantment URL if, for some
inexplicable reason, the briefcase is still in hangk’s. I can’t believe
this, this script is supporting bad moon and standard exploits? Right
in front of a dev?